+++ app/controllers/api/v1/issue_reasons_controller.rb
@@ -1,0 +1,10 @@
+module Api::V1
+  class IssueReasonsController < ApiController
+
+    def index
+      @issue_reasons = IssueReason.all
+      render json: @issue_reasons
+    end
+
+  end
+end
