+++ app/controllers/api/v1/issue_types_controller.rb
@@ -1,0 +1,10 @@
+module Api::V1
+  class IssueTypesController < ApiController
+
+    def index
+      @issue_types = IssueType.all
+      render json: @issue_types
+    end
+
+  end
+end
