+++ app/controllers/api/v1/states_controller.rb
@@ -1,0 +1,10 @@
+module Api::V1
+  class StatesController < ApiController
+
+    def index
+       @states = State.all
+       render json: @states
+    end
+
+  end
+end
