--- app/controllers/issue_reasons_controller.rb
@@ -1,8 +1,0 @@
-class IssueReasonsController  < ApplicationController
-
-  def index
-     @issue_reasons = IssueReason.all
-     render json: @issue_reasons
-  end
-
-end
