--- app/controllers/issue_types_controller.rb
@@ -1,8 +1,0 @@
-class IssueTypesController  < ApplicationController
-
-  def index
-     @issue_types = IssueType.all
-     render json: @issue_types
-  end
-
-end
