--- app/controllers/states_controller.rb
@@ -1,8 +1,0 @@
-class StatesController  < ApplicationController
-
-  def index
-     @states = State.all
-     render json: @states
-  end
-
-end
