+++ config/application.rb
@@ -17,7 +17,7 @@ require "rails/test_unit/railtie"
 Bundler.require(*Rails.groups)
 Dotenv::Railtie.load
 
-module DesafioiApi
+module SacApi
   class Application < Rails::Application
     # Use the responders controller from the responders gem
     config.app_generators.scaffold_controller :responders_controller
