+++ spec/controllers/api/v1/issue_reasons_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe IssueReasonsController, type: :controller do
+RSpec.describe Api::V1::IssueReasonsController, type: :controller do
   before :each do
     FactoryGirl.create_list(:issue_reason, 5)
     get :index
