+++ spec/controllers/api/v1/states_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe StatesController, type: :controller do
+RSpec.describe Api::V1::StatesController, type: :controller do
   before :each do
     FactoryGirl.create_list(:state, 5)
     get :index
